Сделает активное и неактивное состояние страницы
