+++ js/main.js
@@ -1,14 +1,18 @@
-import { adForm, filtersForm } from './constains.js';
 import { getData } from './api.js';
-import { getInactiveState } from './inactive-state.js';
 import { loadMap, createMarkerAdvertisment } from './rendering-map.js';
-import {getActiveState} from './active-state.js';
+import { getInactiveState, getActiveState } from './page-status.js';
+import { adForm, filtersForm } from './constains.js';
 //Задаем странице не активное состояние
 getInactiveState(adForm);
 getInactiveState(filtersForm);
 //Загружаем карту
+console.log('Неактивное состояние страницы');
 loadMap()
+  .then(() => getActiveState(adForm))
   //Если карта загрузилась, то загружаем данные
+  .then(() => {
+    console.log('Активное состояние формы');
+  })
   .then(() => {
     getData()
       //Если данные загрузились, создаем маркеры на основе этих данных
@@ -17,8 +21,16 @@ loadMap()
         createMarkerAdvertisment(data);
         console.log('Данные загруженны');
       })
+      .then(() => {
+        getActiveState(filtersForm);
+        console.log('Ативное состояние фильтров');
+      })
       //Если данные не загрузились, выдать сообщение в консоль
-      .catch(() => console.log('Данные не загружены'));
+      .catch(() => {
+        console.log('Данные не загружены');
+      });
   })
   //Если карта не загрузилась выдать сообщение в консоль
-  .catch(() => console.log('Карта не загружена'));
+  .catch(() => {
+    console.log('Карта не загружена');
+  });
